added uncompress-v4
added logic to check that the commands needed to unpack each file type, are available before trying to uncompress each. also added a function to run just a check of the commands, so that you can test before getting an error. then added variables and logic to store if a check is successful, so that the checks can be skipped if already confirmed.